pyproject.toml
[project]
name = "2026-01-find-restaurant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "geopy>=2.4.1",
    "googlemaps>=4.10.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
